+++ src/CutCounter.cxx
@@ -1,6 +1,6 @@
 #include "CutCounter.hh"
 
-int& CutCounter::operator[](std::string key) 
+double& CutCounter::operator[](std::string key) 
 { 
   bool new_cut = (m_counts.find(key) == m_counts.end()); 
   if (new_cut) { 
@@ -9,11 +9,11 @@ int& CutCounter::operator[](std::string key)
   return m_counts[key]; 
 }
 
-std::vector< std::pair<std::string, int> > CutCounter::get_ordered_cuts() 
+std::vector< std::pair<std::string, double> > CutCounter::get_ordered_cuts() 
   const 
 {
   typedef std::vector<std::string>::const_iterator IdxItr; 
-  std::vector< std::pair<std::string, int> > ordered_cuts; 
+  std::vector< std::pair<std::string, double> > ordered_cuts; 
   for (IdxItr itr = m_cuts.begin(); itr != m_cuts.end(); itr++) { 
     ordered_cuts.push_back(*m_counts.find(*itr));
   }
